Implement terrain mesh rendering from heightmap with color gradient

+++ src/rendering/renderer_3d.rs
@@ -38,7 +38,7 @@ impl egui_wgpu::CallbackTrait for Renderer3D {
         render_pass.set_pipeline(&state.pipeline);
         render_pass.set_bind_group(0, &state.uniform_bind_group, &[]);
         render_pass.set_vertex_buffer(0, state.vertex_buffer.slice(..));
-        render_pass.set_index_buffer(state.index_buffer.slice(..), wgpu::IndexFormat::Uint16);
+        render_pass.set_index_buffer(state.index_buffer.slice(..), wgpu::IndexFormat::Uint32);
         render_pass.draw_indexed(0..state.num_indices, 0, 0..1);
     }
 }
